Support for parsing function definitions
Language parsing now supports python definitions (There is no
support for using them yet)

In order to do this "native" code need to be added to the
Lexer. Unfortunately this means the parser and lexer files needed
to be split. The only client visible side effect of this is that
the SignalFlowV2Visitor is now called SignalFlowV2ParserVisitor.

Also in keeping with the official python.g4, statement was
renamed to exprStatement (as this is really what it was)

Cleaned up generate.sh to reduce C+P code and also added
"make" like steps for each port to check for compilation issues